chore: move backend source code into infra folder
