--- packages/backend/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@aws-sdk-notes-app/backend",
-  "version": "1.0.0",
-  "private": true,
-  "description": "Backend for notes app created using modular AWS SDK for JavaScript",
-  "dependencies": {
-    "@aws-sdk/client-dynamodb": "3.348.0",
-    "@aws-sdk/util-dynamodb": "3.348.0"
-  },
-  "keywords": [
-    "modular",
-    "aws",
-    "JavaScript",
-    "TypeScript",
-    "SDK",
-    "v3"
-  ],
-  "author": "Kamat, Trivikram",
-  "license": "MIT",
-  "devDependencies": {
-    "@types/aws-lambda": "^8.10.64",
-    "@types/node": "^18.11.18",
-    "@typescript-eslint/eslint-plugin": "^5.48.1",
-    "@typescript-eslint/parser": "^5.48.1",
-    "eslint": "^8.31.0",
-    "typescript": "~4.9.4"
-  },
-  "sideEffects": false
-}
