--- packages/backend/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2018",
-    "module": "ES6",
-    "moduleResolution": "node",
-    "esModuleInterop": true
-  },
-  "include": ["src/**/*.ts"]
-}
