+++ packages/infra/package.json
@@ -8,11 +8,14 @@
     "cdk": "tsc && cdk"
   },
   "devDependencies": {
+    "@types/aws-lambda": "^8.10.64",
     "@types/node": "^18.11.18",
     "aws-cdk": "2.134.0",
     "typescript": "~4.9.4"
   },
   "dependencies": {
+    "@aws-sdk/client-dynamodb": "3.348.0",
+    "@aws-sdk/util-dynamodb": "3.348.0",
     "aws-cdk-lib": "2.134.0",
     "constructs": "10.3.0"
   }
